+++ AssemblyLib/Remapper/FilterService.cs
@@ -1,6 +1,7 @@
 ﻿using AsmResolver.DotNet;
 using AssemblyLib.Models;
 using AssemblyLib.ReMapper.Filters;
+using AssemblyLib.Shared;
 using Serilog;
 using SPTarkov.DI.Annotations;
