+++ AssemblyLib/Remapper/MappingController.cs
@@ -4,9 +4,11 @@ using AsmResolver.DotNet;
 using AsmResolver.DotNet.Code.Cil;
 using AsmResolver.PE.DotNet.Cil;
 using AsmResolver.PE.DotNet.Metadata.Tables.Rows;
+using AssemblyLib.Extensions;
 using AssemblyLib.Models;
 using AssemblyLib.Remapper;
 using AssemblyLib.ReMapper.MetaData;
+using AssemblyLib.Shared;
 using AssemblyLib.Utils;
 using Serilog;
 using Serilog.Events;
